# Speed up rabbit tests suites