--- a/deps/rabbit/test/peer_discovery_classic_config_SUITE.erl
+++ b/deps/rabbit/test/peer_discovery_classic_config_SUITE.erl
@@ -15,6 +15,7 @@
     cluster_members_online/2
 ]).
 
+-compile(nowarn_export_all).
 -compile(export_all).
 
 all() ->