--- a/deps/rabbit/test/queue_type_SUITE.erl
+++ b/deps/rabbit/test/queue_type_SUITE.erl
@@ -65,14 +65,6 @@ init_per_group(Group, Config) ->
             ok = rabbit_ct_broker_helpers:rpc(
                    Config2, 0, application, set_env,
                    [rabbit, channel_tick_interval, 100]),
-            %% HACK: the larger cluster sizes benefit for a bit more time
-            %% after clustering before running the tests.
-            case Group of
-                cluster_size_5 ->
-                    timer:sleep(5000);
-                _ ->
-                    ok
-            end,
             Config2
     end.