rabbit_stream_queue_SUITE:
```
BEFORE: time gmake -C deps/rabbit ct-rabbit_stream_queue  7.22s user 5.72s system 2% cpu 8:28.18 total

AFTER time gmake -C deps/rabbit ct-rabbit_stream_queue  27.04s user 8.43s system 10% cpu 5:38.63 total
```